task(refactor): abstract out notification click callback workaround for killed state SDK-1360
