+++ Example/package.json
@@ -13,13 +13,12 @@
     "@react-navigation/drawer": "^5.12.5",
     "@react-navigation/native": "^6.0.6",
     "@react-navigation/native-stack": "^6.2.5",
-    "clevertap-react-native": "git+https://github.com/CleverTap/clevertap-react-native.git#develop",
+    "clevertap-react-native": "^0.8.0",
     "react": "17.0.1",
     "react-native": "0.64.1",
     "react-native-drawer": "^2.5.1",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-pager-view": "^5.1.2",
-    "react-native-reanimated": "^2.3.1",
     "react-native-safe-area-context": "^3.2.0",
     "react-native-screens": "^3.0.0",
     "react-native-tab-view": "^3.0.1",
